refactor(desktop): move remaining domain modules to features/
budget, permissions, phases, plans, skills, worktree moved.
all importers and vi.mock references updated.
