+++ apps/desktop/src/__tests__/store/permission-scope-cta.test.ts
@@ -11,7 +11,7 @@ vi.mock('../../turn', () => ({
 
 const permissionRuleUpsertSpy = vi.fn();
 
-vi.mock('../../permissions', () => ({
+vi.mock('../../features/permissions/permissions', () => ({
   invokePermissionRuleList: vi.fn(async () => []),
   invokePermissionRuleUpsert: (...args: unknown[]) => permissionRuleUpsertSpy(...args),
   invokePermissionAuditInsert: vi.fn(async () => undefined),
@@ -74,7 +74,7 @@ vi.mock('../../routing', () => ({
   })),
 }));
 
-vi.mock('../../budget', () => ({
+vi.mock('../../features/budget/budget', () => ({
   invokeBudgetRuleList: vi.fn(async () => []),
   invokeBudgetRuleUpsert: vi.fn(),
   invokeBudgetRuleDelete: vi.fn(),
@@ -85,7 +85,7 @@ vi.mock('../../budget', () => ({
   invokeCheckProviderBudget: vi.fn(),
 }));
 
-vi.mock('../../skills', () => ({
+vi.mock('../../features/skills/skills', () => ({
   invokeSkillList: vi.fn(async () => []),
   invokeSkillUpsert: vi.fn(),
   invokeSkillDelete: vi.fn(),
@@ -93,7 +93,7 @@ vi.mock('../../skills', () => ({
   resolveSkillInvocation: vi.fn(),
 }));
 
-vi.mock('../../phases', () => ({
+vi.mock('../../features/phases/phases', () => ({
   invokePhaseTemplateList: vi.fn(async () => []),
   invokePhaseTemplateUpsert: vi.fn(),
   invokePhaseTemplateDelete: vi.fn(),
@@ -102,7 +102,7 @@ vi.mock('../../phases', () => ({
   invokePhaseRunUpdateStatus: vi.fn(),
 }));
 
-vi.mock('../../worktree', () => ({
+vi.mock('../../features/worktree/worktree', () => ({
   createWorktree: vi.fn(),
   removeWorktree: vi.fn(),
 }));
