+++ apps/desktop/src/__tests__/store/unknown-payload.test.ts
@@ -9,7 +9,7 @@ vi.mock('../../turn', () => ({
   isAuthErrorMessage: () => false,
 }));
 
-vi.mock('../../permissions', () => ({
+vi.mock('../../features/permissions/permissions', () => ({
   invokePermissionRuleList: vi.fn(async () => []),
   invokePermissionAuditInsert: vi.fn(async () => undefined),
   invokeAuditRetryEnqueue: vi.fn(async () => undefined),
@@ -71,7 +71,7 @@ vi.mock('../../routing', () => ({
   })),
 }));
 
-vi.mock('../../budget', () => ({
+vi.mock('../../features/budget/budget', () => ({
   invokeBudgetRuleList: vi.fn(async () => []),
   invokeBudgetRuleUpsert: vi.fn(),
   invokeBudgetRuleDelete: vi.fn(),
@@ -82,7 +82,7 @@ vi.mock('../../budget', () => ({
   invokeCheckProviderBudget: vi.fn(),
 }));
 
-vi.mock('../../skills', () => ({
+vi.mock('../../features/skills/skills', () => ({
   invokeSkillList: vi.fn(async () => []),
   invokeSkillUpsert: vi.fn(),
   invokeSkillDelete: vi.fn(),
@@ -90,7 +90,7 @@ vi.mock('../../skills', () => ({
   resolveSkillInvocation: vi.fn(),
 }));
 
-vi.mock('../../phases', () => ({
+vi.mock('../../features/phases/phases', () => ({
   invokePhaseTemplateList: vi.fn(async () => []),
   invokePhaseTemplateUpsert: vi.fn(),
   invokePhaseTemplateDelete: vi.fn(),
@@ -99,7 +99,7 @@ vi.mock('../../phases', () => ({
   invokePhaseRunUpdateStatus: vi.fn(),
 }));
 
-vi.mock('../../worktree', () => ({
+vi.mock('../../features/worktree/worktree', () => ({
   createWorktree: vi.fn(),
   removeWorktree: vi.fn(),
 }));
