+++ apps/desktop/src/components/SessionSettingsDialog.tsx
@@ -15,7 +15,7 @@ import { formatError } from '../shared/lib/errors';
 import { useAppStore } from '../store';
 import { SESSION_FEATURES } from '../shared/lib/features';
 import { parseCap } from '../shared/lib/parse-cap';
-import { listLocalBranches, type LocalBranchInfo } from '../worktree';
+import { listLocalBranches, type LocalBranchInfo } from '../features/worktree/worktree';
 import { useToast } from './Toast';
 
 interface SessionSettingsDialogProps {
